trying to compile all

+++ Client/src/Box2D/box2dManager.cpp
@@ -106,7 +106,7 @@ void BoxSimulator::run() {
         worm->SetLinearVelocity(vel);  // seteo la nueva velocidad
         world->Step(timeStep, velocityIterations, positionIterations); //simulo un paso con la info actual
         b2Vec2 pos = worm->GetPosition(); //consigo la pos
-        std::vector<int> positions = {int(pos.x*100.0f), int(pos.y*100.0f)};
+        std::vector<float> positions = {float(pos.x*100.0f), float(pos.y*100.0f)};
         outgoing.push(positions); //paso la pos
     }
 }
